+++ resources/views/v2/import/004-configure/index.blade.php
@@ -55,6 +55,8 @@
                 <input type="hidden" name="nordigen_max_days" value="{{ $configuration->getNordigenMaxDays() }}"/>
                 <input type="hidden" name="nordigen_bank" value="{{ $configuration->getNordigenBank() }}"/>
                 <input type="hidden" name="nordigen_requisitions" value="{{ json_encode($configuration->getNordigenRequisitions()) }}"/>
+                <!-- these values are used by SimpleFIN and must be preserved -->
+                <input type="hidden" name="access_token" value="{{ $configuration->getAccessToken() }}"/>
 
                 @if('nordigen' === $flow || 'spectre' === $flow || 'simplefin' === $flow)
                     <input type="hidden" name="ignore_duplicate_transactions" value="1"/>